requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
python-dotenv>=1.0.0
openai>=1.0.0
pydantic==1.10.13
streamlit==1.32.0 